% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mlogit.R
\name{construct.v}
\alias{construct.v}
\title{Split Names of Vectors into N-vectors
This function is used to organize how variables are spread
across the list of formulas}
\usage{
construct.v(constraints, ndim)
}
\arguments{
\item{constraints}{a constraints object}

\item{ndim}{an integer specifying the number of dimensions}
}
\value{
a list of character-vectors
}
\description{
Split Names of Vectors into N-vectors
This function is used to organize how variables are spread
across the list of formulas
}
