% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mlogit.R
\name{pv.mlogit}
\alias{pv.mlogit}
\title{Simulate Predicted Values}
\usage{
pv.mlogit(fitted, ev)
}
\arguments{
\item{fitted}{a fitted model object}

\item{ev}{the simulated expected values}
}
\value{
a vector of simulated values
}
\description{
Simulate Predicted Values
}
