% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-blogit.R
\docType{class}
\name{Zelig-blogit-class}
\alias{Zelig-blogit-class}
\alias{zblogit}
\title{Bivariate Logistic Regression for Two Dichotomous Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zeligchoice-blogit.html}
}

