% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model-mlogit.R
\docType{class}
\name{Zelig-mlogit-class}
\alias{Zelig-mlogit-class}
\alias{zmlogit}
\title{Multinomial Logistic Regression for Dependent Variables with Unordered Categorical Values}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zeligchoice-mlogit.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, ..., weights = NULL, by)}}{The zelig command estimates a variety of statistical models}
}}

