% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ologit.R
\docType{class}
\name{Zelig-ologit-class}
\alias{Zelig-ologit-class}
\alias{zologit}
\title{Ordinal Logistic Regression for Ordered Categorical Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zeligchoice-ologit.html}
}

