% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model-oprobit.R
\docType{class}
\name{Zelig-oprobit-class}
\alias{Zelig-oprobit-class}
\alias{zoprobit}
\title{Ordinal Probit Regression for Ordered Categorical Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zeligchoice-oprobit.html}
}
\section{Methods}{

\describe{
\item{\code{test(z, data)}}{Monte Carlo Unit Test}
}}

