% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bbinchoice.R
\docType{class}
\name{Zelig-bbinchoice-class}
\alias{Zelig-bbinchoice-class}
\alias{zbbinchoice}
\title{Bivariate Binary Choice object for inheritance across models in ZeligChoice}
\description{
Bivariate Binary Choice object for inheritance across models in ZeligChoice
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

