% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bprobit.R
\docType{class}
\name{Zelig-bprobit-class}
\alias{Zelig-bprobit-class}
\alias{zbprobit}
\title{Bivariate Probit Regression for Two Dichotomous Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zeligchoice-bprobit.html}
}

