% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mlogit.R
\docType{class}
\name{Zelig-mlogit-class}
\alias{Zelig-mlogit-class}
\alias{zmlogit}
\title{Multinomial Logistic Regression for Dependent Variables with Unordered Categorical Values}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zeligchoice_mlogit.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

