% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-obinchoice.R
\docType{class}
\name{Zelig-obinchoice-class}
\alias{Zelig-obinchoice-class}
\alias{zobinchoice}
\title{Ordered Choice object for inheritance across models in ZeligChoice}
\description{
Ordered Choice object for inheritance across models in ZeligChoice
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

