% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model-ologit.R
\docType{class}
\name{Zelig-ologit-class}
\alias{Zelig-ologit-class}
\alias{zologit}
\title{Ordinal Logistic Regression for Ordered Categorical Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zeligchoice-ologit.html}
}
\section{Methods}{

\describe{
\item{\code{test(z, data)}}{Monte Carlo Unit Test}
}}

